WebAppStreamlit/requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
ansi2html==1.9.2
anyio==4.7.0
astunparse==1.6.3
attrs==24.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dash==2.11.0
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_uploader==0.6.1
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
faiss-cpu==1.9.0.post1
fake-useragent==1.5.1
filelock==3.16.1
filetype==1.2.0
flake8==6.1.0
Flask==2.2.5
flatbuffers==24.12.23
fonttools==4.55.3
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.13.1
grpcio==1.68.1
gunicorn==21.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
kiwisolver==1.4.7
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.28
langchain-groq==0.2.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.4
langdetect==1.0.9
langsmith==0.1.147
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.20.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.0
olefile==0.47
openai==1.58.1
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.13
packaging==24.2
pandas==2.1.3
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pip==24.3.1
platformdirs==4.3.6
plotly==5.15.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.7.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
pypdf==5.0.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
referencing==0.35.1
regex==2024.11.6
reportlab==4.0.4
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.3.1
sentencepiece==0.2.0
setuptools==65.5.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
streamlit==1.38.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tk==0.1.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
unstructured==0.15.13
unstructured-client==0.28.1
uritemplate==4.1.1
urllib3==2.2.3
watchdog==4.0.2
Werkzeug==2.2.3
wheel==0.45.1
wrapt==1.17.0
xlrd==2.0.1
yarl==1.18.3
youtube-transcript-api==0.6.2
zipp==3.21.0
